+++ internal/helper.go
@@ -9,6 +9,8 @@ import (
 	"fmt"
 )
 
+// Bprintf formats according to a format specifier and append it to passed buffer.
+// It returns passed buffer with added text.
 func Bprintf(buf []byte, format string, a ...interface{}) []byte {
 	b := bytes.NewBuffer(buf)
 	// fmt.Fprintf calls b.Write, which never returns error