+++ CHANGELOG.md
@@ -2,9 +2,24 @@
 
 ## [Unreleased]
 
+## [0.2.0] - 2022-02-24
+### Added
+- Introduced new constants `size.ObjectKeyValue` and `size.ObjectKeyUnit`.
+- Missing doc comments.
+
+### Changed
+- Renamed `size.NewSize` to `size.New`.
+- Optimized `size.unmarshalJSONObject` implementation.
+
+### Removed
+- Variable `sem.Zero`.
+  - Use `sem.Ver{}` instead.
+- Name of `date.filterNo.Contains` receiver.
+
 ## [0.1.0] - 2022-02-22
 ### Added
 - First release of util.
 
-[Unreleased]: https://github.com/livesport-tv/util/compare/v0.1.0...master
+[Unreleased]: https://github.com/livesport-tv/util/compare/v0.2.0...master
+[0.2.0]: https://github.com/livesport-tv/util/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/livesport-tv/util/releases/tag/v0.1.0